style: fix + button centering and add fake profile component
